move randomizers into utils.go
