+++ pkg/hintrunner/hint.go
@@ -1,10 +1,8 @@
 package hintrunner
 
 import (
-	"encoding/binary"
 	"fmt"
 	"math/big"
-	"math/rand"
 	"sort"
 
 	"github.com/holiman/uint256"
@@ -1333,27 +1331,4 @@ func (hint *RandomEcPoint) Execute(vm *VM.VirtualMachine) error {
 	}
 
 	return nil
-}
-
-func randomFeltElement(rand *rand.Rand) f.Element {
-	b := [32]byte{}
-	binary.BigEndian.PutUint64(b[24:32], rand.Uint64())
-	binary.BigEndian.PutUint64(b[16:24], rand.Uint64())
-	binary.BigEndian.PutUint64(b[8:16], rand.Uint64())
-	//Limit to 59 bits so at max we have a 251 bit number
-	binary.BigEndian.PutUint64(b[0:8], rand.Uint64()>>5)
-	f, _ := f.BigEndian.Element(&b)
-	return f
-}
-
-func randomFeltElementU128(rand *rand.Rand) f.Element {
-	b := [32]byte{}
-	binary.BigEndian.PutUint64(b[24:32], rand.Uint64())
-	binary.BigEndian.PutUint64(b[16:24], rand.Uint64())
-	f, _ := f.BigEndian.Element(&b)
-	return f
-}
-
-func defaultRandGenerator() *rand.Rand {
-	return rand.New(rand.NewSource(0))
 }
